+++ src/models/Game/gameStatus.ts
@@ -1,9 +1,6 @@
 enum GameStatus {
     WaitingForPlayers = 0,
     Playing = 1,
-    AssassinTurn=2,
-    GoodWin = 3,
-    EvilWin = 4,
     Ended = -1
 }
