+++ src/models/Players/player.ts
@@ -3,15 +3,3 @@ export interface Player {
     name: string;
     role: string;
 }
-
-export class EmptyPlayer implements Player {
-    id: string;
-    name: string;
-    role: string;
-
-    constructor() {
-        this.id = "";
-        this.name = "";
-        this.role = "";
-    }
-}
