+++ src/scenes/GameStart/index.tsx
@@ -1,26 +1,20 @@
 import React, { useState, useEffect } from 'react';
-import { Row, Col, List, Button } from 'antd';
+import { Row, Col, Button, List } from 'antd';
 import { inject, observer } from 'mobx-react';
 import Stores from '../../stores/storeIdentifier';
 import './index.less';
+// import signalRAspNetCoreHelper from '../../lib/signalRAspNetCoreHelper';
 import { HubConnectionBuilder } from '@aspnet/signalr';
 import { Player } from '../../models/Players/player';
-import AppConsts from '../../lib/appconst';
-import { useHistory } from 'react-router-dom'
-
-
+// import GameStore from '../../stores/gameStore';
 declare var abp: any;
 
+// interface GameStartProps {
+//     gameStore: GameStore;
+// }
 const GameStart = (props: any) => {
+    // const [socket, setSocket] = useState<HubConnection>();
     const [game, setGame] = useState(props.gameStore.currentGame);
-    const history = useHistory();
-
-
-    useEffect(() => {
-        (async () => {
-            await createGame();
-        })();
-    }, [])
 
     useEffect(() => {
         setGame(props.gameStore.currentGame)
@@ -33,7 +27,7 @@ const GameStart = (props: any) => {
 
     const initSocket = async () => {
         const connect = new HubConnectionBuilder()
-            .withUrl(AppConsts.remoteServiceBaseUrl + '/gameHub')
+            .withUrl('https://multiplayeravalon.azurewebsites.net/gameHub')
             .build()
 
         try {
@@ -45,6 +39,35 @@ const GameStart = (props: any) => {
         connect.on("GameUpdated", function() {
             props.gameStore.get(props.gameStore.currentGame.id);
         })
+
+        // setSocket(connect)
+    }
+
+    let view;
+    if(game.joinCode) {
+        view = 
+            <>
+                <h3>Join the game on your device with this code:</h3>
+                <h1>{game.joinCode}</h1>
+                <h3>Current Players:</h3>
+                <List
+                    grid={{
+                        gutter: 16,
+                        xs: 1,
+                        sm: 2,
+                        md: 4,
+                        lg: 4,
+                        xl: 6,
+                        xxl: 3}}
+                    dataSource={game.players} 
+                    renderItem={(player: Player) => (
+                        <List.Item>
+                            {player.name}
+                        </List.Item>
+                    )}/>
+            </>
+    } else {
+        view = <Row><Button onClick={createGame}>Create Game</Button></Row>
     }
 
     return(
@@ -52,28 +75,8 @@ const GameStart = (props: any) => {
             <Col sm={24} lg={16} xl={10}>
                 <Row justify="center" className="createGameView">
                     <Col>
-                    <h3>Join the game on your device with this code:</h3>
-                    <h1>{game.joinCode}</h1>
-                    <h3>Current Players:</h3>
-                    <List
-                        grid={{
-                            gutter: 16,
-                            xs: 1,
-                            sm: 2,
-                            md: 4,
-                            lg: 4,
-                            xl: 6,
-                            xxl: 3}}
-                        dataSource={game.players} 
-                        renderItem={(player: Player) => (
-                            <List.Item>
-                                {player.name}
-                            </List.Item>
-                        )}/>
+                        {view}
                     </Col>
-                </Row>
-                <Row>
-                    <Button onClick={() => history.push("/startGame")}>Start Game</Button>
                 </Row>
             </Col>
         </Row>
