--- src/scenes/PlayGame/components/WaitingForPlayers/index.less
@@ -1,5 +1,0 @@
-.waitingForPlayers {
-    .header {
-        text-align: center;
-    }
-}
