--- src/services/dto/gamePlayerDto.ts
@@ -1,7 +1,0 @@
-import { Game } from '../../models/Game/game';
-import { Player } from '../../models/Players/player';
-
-export interface GamePlayerDto {
-    game: Game;
-    player: Player;
-}
