+++ src/stores/storeIdentifier.ts
@@ -1,5 +1,4 @@
 export default class Stores {
   static GameStore: string = 'gameStore';
   static SessionStore: string = 'sessionStore';
-  static PlayerStore: string = 'playerStore';
 }
